requirements.txt
Flask==3.0.3
Pillow==12.0.0
requests==2.32.3
gunicorn==22.0.0